Changes to js file structure
Changes to js file structure for convenience
