+++ src/auth/src/main/resources/static/js/data/routes/base-route.js
@@ -1,0 +1,5 @@
+"use strict";
+
+const BASE_ROUTE = "http://localhost:8080";
+
+export default BASE_ROUTE;
