+++ src/auth/src/main/resources/static/js/data/routes/routes.js
