+++ src/auth/src/main/resources/static/js/utils/fetch/data/fetch-data.js
@@ -1,0 +1,7 @@
+"use strict";
+
+import FETCH_ERRORS from "./fetch-errors.js";
+import FETCH_HEADERS from "./fetch-headers.js";
+import FETCH_METHODS from "./fetch-methods.js";
+
+export { FETCH_ERRORS, FETCH_HEADERS, FETCH_METHODS };
