+++ src/auth/src/main/resources/static/js/utils/fetch/data/fetch-errors.js
@@ -1,0 +1,10 @@
+"use strict";
+
+const FETCH_ERRORS = {
+	"response-error": "There was a problem with receiving response from server",
+	"already-exists": "Username is already taken or email is already in use",
+	"illegal-field": "There was a problem with data you sent to the server",
+	"login-error": "Incorrect username or password when logging in"
+};
+
+export default FETCH_ERRORS;
