+++ src/auth/src/main/resources/static/js/utils/fetch/data/fetch-methods.js
@@ -1,0 +1,5 @@
+"use strict";
+
+const FETCH_METHODS = { post: "POST", get: "GET" };
+
+export default FETCH_METHODS;
