+++ src/auth/src/main/resources/static/js/utils/fetch/headers/headers.js
