+++ src/auth/src/main/resources/static/js/utils/parser/dom/dom-getters.js
@@ -1,0 +1,8 @@
+"use strict";
+
+const DOM_DATA_GETTERS = {
+	name: (name) => document.getElementsByName(name)[0].value,
+	id: (id) => document.getElementById(id),
+};
+
+export default DOM_DATA_GETTERS;
