+++ src/auth/src/main/resources/static/js/utils/parser/dom/dom-parser.js
@@ -1,0 +1,9 @@
+"use strict";
+
+const parseDomData = (method, ...args) => {
+	const userData = {};
+	for (const element of args) userData[element] = method(element);
+	return userData;
+};
+
+export default parseDomData;
