+++ src/auth/src/main/resources/static/js/utils/validators/data/validator-expression.js
@@ -7,9 +7,4 @@ const VALIDATOR_EXPRESSIONS = {
 	empty: /\A\s*\z/,
 }
 
-const validate = (str, validator) => {
-	if (str == null) return false;
-	return validator.test(String(str));
-}
-
-export { VALIDATOR_EXPRESSIONS, validate };
+export default VALIDATOR_EXPRESSIONS;
