+++ src/auth/src/main/resources/static/js/utils/validators/fields/validate-fields.js
