+++ src/auth/src/main/resources/static/js/utils/validators/validator-implementations/generic/validate.js
@@ -1,0 +1,8 @@
+"use strict";
+
+const validate = (str, validator) => {
+	if (str == null) return false;
+	return validator.test(String(str));
+}
+
+export default validate;
