+++ src/auth/src/main/resources/static/js/utils/validators/validator-implementations/implementations/email-validator.js
