+++ src/auth/src/main/resources/static/js/utils/validators/validator-implementations/implementations/empty-validator.js
